requirements.txt
appnope==0.1.0
asgiref==3.2.7
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
docopt==0.6.2
entrypoints==0.3
Flask==1.0.3
gunicorn==20.0.4
idna==2.9
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mongo==0.2.0
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.16.2
panda==0.3.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pipenv==2018.11.26
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.5
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pygal==2.4.0
Pygments==2.3.1
pymongo==3.8.0
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pythonanywhere==0.9.1
pytz==2019.3
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.23.0
schema==0.7.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.1
sqlparse==0.3.1
sympy==1.4
tabulate==0.8.7
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
urllib3==1.25.8
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==5.0.1
widgetsnbextension==3.4.2
